Refactor: Rename shared library to padawankmp
