+++ .gitignore
@@ -11,7 +11,7 @@
 /androidApp/debug
 /androidApp/gradle.properties
 
-/shared/build
+/padawankmp/build
 
 xcuserdata/
